requirements.txt
# python3==3.10.2
pymongo
# blinker==1.7.0
certifi
# charset-normalizer==3.3.2
# click==8.1.7
# docker==7.0.0
et-xmlfile
flask
python-dotenv
# Jinja2==3.1.3
joblib
# MarkupSafe==2.1.3
numpy
# openpyxl==3.1.2
packaging
pandas
pip
python-dateutil
# pytz==2023.3.post1
requests
scikit-learn
scipy
# setuptools==68.2.2
# six==1.16.0
tzdata
urllib3
# werkzeug==3.0.1
# wheel==0.41.2
gunicorn
